+++ Source/nightmare/world/ambient_animals.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace devilution {
+namespace nightmare {
+
+void InitTristramAmbientAnimals();
+void UpdateTristramAmbientAnimals();
+
+} // namespace nightmare
+} // namespace devilution